added css and refactored code for TicTacToe
