+++ src/TicTacToe/TicTacToeStyling.js
@@ -2,6 +2,7 @@ import {createGlobalStyle} from "styled-components"
 
 const TicTacToeStyle = createGlobalStyle`
 @import url('https://fonts.googleapis.com/css2?family=Sono:wght@400;700&display=swap');
+@import url('https://fonts.googleapis.com/css?family=Rancho&effect=fire|neon');
 html, body{
   min-height: 100vh;
   padding: 0;
@@ -61,5 +62,10 @@ table tr td:last-child
 {
   box-shadow:-8px 0px 5px -5px cornsilk;
 }
+td:hover
+{
+background: paleturquoise;
+}
+
 `
 export default TicTacToeStyle
